Rename RadioBeta.py to RadioAlpha2.py
